# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Swagger ToDo Manager
  description: REST APIs for the interaction with the ToDo Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
  /api/tasks/public:
    get:
      tags:
        - tasks/public
      summary: Retrieve the public tasks
      description: >-
        The public tasks (i.e., the tasks that are visible for all the users of the service are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicTasks
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  next:
                    type: string
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  /api/tasks:
    post:
      tags:
        - tasks
      security:
        - cookieAuth: []
      summary: Create a new task
      description: A new task is created by the authenticated user. This API does not allow to assign the task to users.
      operationId: addTask
      requestBody:
        description: Task object that needs to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
  '/api/tasks/{taskId}':
    put:
      tags:
        - 'tasks/{taskId}'
      security:
        - cookieAuth: []
      summary: Update a task
      description: The task with ID taskId is updated. This operation does not allow to mark a task as completed. This operation can be performed only by the owner.
        This API does not allow to update the lists of assignees, active users or users that have completed the task.
        If the completers property is modified, one of the following happens --> 
        (1) completers number is decreased --> if the actual number of users that completed the task is now greater than or equal to the "completers" property, the task is considered completed, and its flag is set to 1
        (2) completers number is increased --> if the actual number of users that completed the task is not sufficient anymore to consider the entire task as completed, the completed flag is set to 0
      operationId: updateSingleTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated task object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    get:
      tags:
        - 'tasks/{taskId}'
      security:
        - cookieAuth: []
      summary: Retreve a task
      description: The task with ID taskId is retrieved. This operation can be performed on the task if at least one of the following three conditions is satisfied. 1) The task is public. 2) The user who performs the operation is the task's owner. 3) The user who performs the operation is a task's assignee. This API does not retrieve the content of the list of assignees, active users and who completed the task.
      operationId: getSingleTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - 'tasks/{taskId}'
      security:
        - cookieAuth: []
      summary: Delete a task
      description: The task with ID taskId is deleted. This operation can only be performed by the owner.
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  /api/users:
    get:
      tags:
        - users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: >-
        The available information (passwords excluded) about all the
        users is retrieved.
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not Found
          content: {}
...
  '/api/tasks/{taskId}/completion':
    put:
      tags:
        - 'tasks/{taskId}/completion'
      security:
        - cookieAuth: []
      summary: Complete a task
      description: >-
        The task with ID taskId is marked as completed for the logged in user. 
        The user must be an assignee for the task. 
        If the user already completed the task, a conflict error is returned.
        If the user was active on the task, after completing the task, the user will not be anymore an active user on the task.
        If, after the completion, the number of completers is enough to consider the entire task completed, the completed flag will be set to 1.
      operationId: completeTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    get:
      tags:
        - 'tasks/{taskId}/completion'
      security:
        - cookieAuth: []
      summary: Retrieve the list of users that completed the task
      description: >-
        Retrieve the list of users that completed the task. This operation can only be performed by the owner of the task.
      operationId: getTaskCompleters
      parameters:
        - name: taskId
          in: path
          description: ID of the task
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/users/{userId}/tasks/created':
    get:
      tags:
        - users/{userId}/tasks/created
      security:
        - cookieAuth: []
      summary: Retrieve the tasks created by the user with ID userId
      description: >-
        This operation allows an authenticated user to retrieve the tasks that she has created. 
        This API does not retrieve the content of the list of assignees, active users and who completed the task.
        A pagination mechanism is implemented to limit the size of messages.
      operationId: getUserTasks
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  next:
                    type: string
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/users/{userId}/tasks/assigned':
    get:
      tags:
        - users/{userId}/tasks/assigned
      security:
        - cookieAuth: []
      summary: Retrieve the tasks assigned to the user with ID userId
      description: >-
        This operation allows an authenticated user to retrieve the tasks that has been assigned to her.
        This API does not retrieve the content of the list of assignees, active users and who completed the task.
        A pagination mechanism is implemented to limit the size of messages.
      operationId: getAssignedTasks
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  next:
                    type: string
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
components:
  schemas:
    Task:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the task
        description:
          type: string
          maxLength: 160
          description: textual task description
        important:
          type: boolean
          description: >-
            this property is set to true if the task is marked as important,
            false otherwise
          default: false
        private:
          type: boolean
          description: >-
            this property is set to true if the task is marked as private, false
            otherwise
          default: true
        project:
          type: string
          description: project in which the task is inserted
        deadline:
          type: string
          format: date-time
          description: the due date and hour of the task
        completed:
          type: boolean
          description: >-
            this property is set to true if the task is marked as completed,
            false otherwise
          default: false
        owner: 
          $ref: '#/components/schemas/User'
        assignees:
          type: array
          description: >-
            the list of users that has this task assigned
          items:
            $ref: '#/components/schemas/User'
        active:
          type: array
          description: >-
            the list of users that is currently working on this task
          items:
            $ref: '#/components/schemas/User'
        completer:
          type: array
          description: >-
            the list of users that have already completed this task
          items:
            $ref: '#/components/schemas/User'
        completers:
          type: integer
          description: >-
            minimum number of assegnee that must complete the task in order to
            consider it completed
          default: 1
        self:
          type: string
          format: uri-reference
          description: self link
    User:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          format: email
          description: email address of the user
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: password of the user
        self:
          type: string
          format: uri-reference
          description: self link
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid